--- Assignment/Directory.Build.props
@@ -1,10 +1,0 @@
-<Project>
-<PropertyGroup>
-    <TargetFramework>net9.0</TargetFramework>
-	<LangVersion>latest</LangVersion>
-    <Nullable>enable</Nullable>
-    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-    <EnableNETAnalyzers>true</EnableNETAnalyzers>
-    <CodeAnalysisTreatWarningsAsErrors>true</CodeAnalysisTreatWarningsAsErrors>
-  </PropertyGroup>
-</Project>
